refactored decoding of 'grid' images
